composite: allow filters to be added on encode path
Removes assertion preventing matchers from inserting a filter on the response path.

It appears like this behavior used to cause a crash or be validated against, but it doesn't appear to still be the case.

This is very useful for end-users who would like to, for example, mutate the response body for a filter when a request header is specified and the content-type matches, which could be especially expensive if the executed filter is ext_proc.

Tested via integration tests (added in this commit).

Signed-off-by: Matt Leon <mattleon@google.com>