Fixes grpc-web filter content length handling
Fixes #40097

This fixes a bug in the grpc web filter where if an upstream grpc server sends
a response with the content length, the grpc web filter modifies the body
without updating or removing the content length. The result is that some
clients truncate the body, other clients throw a protocol error, and if the
downstream protocol is HTTP/1.1, there's a potential for HTTP response
splitting.

Signed-off-by: James Roper <james@jazzy.id.au>