maxmind geoip provider: add gauge to track database creation timestamp
This change adds a gauge metric to track the build epoch of a Maxmind
database file, see https://maxmind.github.io/MaxMind-DB/#build_epoch.
This value is part of the MMDB file formats metadata field and according
to docs should always be present
https://maxmind.github.io/libmaxminddb.

The main usage of this metric would be to track the freshness of the
database files that the Envoy process has loaded into memory.

Signed-off-by: Alex Sundström <asundstrom@spotify.com>