chore: rename tasks_example.json (#40844)
In most cases, we should not submit `tasks.json` to the
code base because in actual development we need to customize a lot (e.g
more tasks or environment variables). Rename `tasks.json` to
`tasks_example.json` instead of deleting it, in case someone want to use
it as an example.

Risk Level: low
Testing: n/a
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a
Signed-off-by: zirain <zirain2009@gmail.com>