composite: allow filters to be added on encode path (#40800)
* Commit Message: composite: allow filters to be added on encode path
* Additional Description:
Removes assertion preventing matchers from inserting a filter on the
response path.

It appears like this behavior used to cause a crash or be validated
against, but it doesn't appear to still be the case.

This is very useful for end-users who would like to, for example, mutate
the response body for a filter when a request header is specified and
the content-type matches, which could be especially expensive if the
executed filter is ext_proc.

* Risk Level: medium (new feature in existing filter)
* Testing: integration. Also tested locally with an end-to-end test with
ext_proc.
* Docs Changes: Removed stipulation that matching can only rely on
request headers.
* Release Notes: Added note that composite filter can be configured to
match outside of request headers.
* Platform Specific Features: None.

Signed-off-by: Matt Leon <mattleon@google.com>