netns: Fix incorrect error status in listener manager
When creating a listener socket in another network namespace, a nested
`absl::StatusOr<>` is returned. The outer status shows the result of the
attempt to switch network namespaces and the inner status is the result
of the creation of a listener socket.

This patch fixes a bug where we checked the inner status when we should
have been checking the outer status. This results in a segfault because
the outer status is dereferenced without first checking if it was OK.

We now check the correct results and additional comments have been added
to improve clarity of this code.

Signed-off-by: Tony Allen <txallen@google.com>