lb: add a new field to use host weight in zone aware lb (#39803)
Additional Description:
Currently, zone-aware load balancing calculates the percentage based on
the number of local and upstream hosts. However, in some scenarios, we
also want to account for the weight assigned to each host. To support
this, I am introducing a new field in ZoneAwareLbConfig. When this field
is enabled, the percentage calculation will consider each host's weight
in addition to the host count.

Testing:
1. unit test
2. I also test the functionality on my private hosts.

---------

Signed-off-by: cqi1217 <caiqi1217@gmail.com>