tls: future-proof UnsupportedCurveEcdsaCert's test expectation
Envoy has logic to reject P-224, but newer versions of BoringSSL reject
it in `SSL_CTX` before Envoy's logic runs. Fix the test expectation to
accept both paths.

(It was already not possible to use a P-224 TLS key in BoringSSL because
we didn't recognize the P-224 NamedCurve codepoint in TLS 1.2, and there
is no P-224 SignatureScheme codepoint in TLS 1.3. BoringSSL just didn't
notice in SSL_CTX_use_certificate before.)

Signed-off-by: David Benjamin <davidben@google.com>