grpc_web: fixes response content length handling (#40804)
Fixes #40097

This fixes a bug in the grpc web filter where if an upstream grpc server
sends a response with the content length, the grpc web filter modifies
the body without updating or removing the content length. The result is
that some clients truncate the body, other clients throw a protocol
error, and if the downstream protocol is HTTP/1.1, there's a potential
for HTTP response splitting.

I have attempted to run the tests, but multiple attempts to run the
tests in docker have caused my machine to go to 100% CPU for multiple
hours before the machine crashes. Running the tests outside of docker
seem to have dependency issues.

Signed-off-by: James Roper <james@jazzy.id.au>