tls: fix no_extension_cert.pem
no_extension_cert.pem was generated by an OpenSSL impacted by
https://github.com/openssl/openssl/issues/28397. This input is not
actually a valid certificate and future versions of BoringSSL may reject
it as invalid.

Refresh it with an OpenSSL with the fix applied. As of writing the fix
has been approved but not yet merged. Because of this, I've made the
generation script double-check the output and error if it was wrong and,
to not interfere with other runs of the script, disabled regeneration of
it for now.

An actual no-extension certificate also makes for a better test because
the extension list will be null. In the certificate Envoy was previously
testing against, the extension list was non-null but empty.

Signed-off-by: David Benjamin <davidben@google.com>