tls: fix no_extension_cert.pem (#40990)
Commit Message:
no_extension_cert.pem was generated by an OpenSSL impacted by
https://github.com/openssl/openssl/issues/28397. This input is not
actually a valid certificate and future versions of BoringSSL may reject
it as invalid.

Refresh it with an OpenSSL with the fix applied. As of writing the fix
has been approved but not yet merged. Because of this, I've made the
generation script double-check the output and error if it was wrong and,
to not interfere with other runs of the script, disabled regeneration of
it for now.

An actual no-extension certificate also makes for a better test because
the extension list will be null. In the certificate Envoy was previously
testing against, the extension list was non-null but empty.

Additional Description:
Risk Level: none, this is a test-only change
Testing: run unit tests, this test-only change
Docs Changes: no, this is a test only change
Release Notes: no, this is a test only change
Platform Specific Features: no, this is a test only change
[Optional Runtime guard:] no, this is a test only change
[Optional Fixes #Issue] no, this is a test only change
[Optional Fixes commit #PR or SHA] no, this is a test only change
[Optional Deprecated:] no, this is a test only change
[Optional [API
Considerations](https://github.com/envoyproxy/envoy/blob/main/api/review_checklist.md):]
no, this is a test only change

---------

Signed-off-by: David Benjamin <davidben@google.com>