Const-correct some uses of X509 API
Many of these APIs should be returning const pointers, but return
non-const because legacy callers store them in the wrong variables. Fix
some of these callers. Also in some cases exposed structs mean the
compiler doesn't check const-ness, but it's still good to tag variables
as const.

In doing so, this fixes a deeper bug in one of the tests: there is no
guarantee that ASN1_STRINGs are NUL-terminated, yet this code was
assuming they were. See CVE-2021-3712. Fortunately, this was just in
test code, or Envoy would have its own instance of that CVE.

(ASN1_STRING_get0_data is the same as ASN1_STRING_data except it acts
on a const pointer.)

Signed-off-by: David Benjamin <davidben@google.com>