xds-federation: introduce OD-CDS xds-federation support (#40988)
Commit Message: xds-federation: introduce OD-CDS xds-federation support
Additional Description:

This PR introduces OD-CDS xds-Federation support. The implementation
uses a new OD-CDS subscription management on the main thread, that is
simpler than the former (issues subscriptions directly using the
xDS-manager subscription function), and doesn't need to maintain the
state of whether the first request was sent or not.
It does so by allowing each subscription be a singleton subscription on
the component side (the gRPC-mux to the xds-Federation server will
ensure that a single subscription message is sent to the server). This
approach handles some bugs that were detected for the non-xDS-TP based
OD-CDS implementation.
The plan is first to use this implementation for xDS-TP based resources,
and when it is validated in production, move the ADS and the
specific-config-source paths to use the new implementation/model.

Apologies for the large PR, but the number of tests that was needed here
is quite large.

Risk Level: low - only for xDS-TP based config sources and it is guarded
by the
`envoy.reloadable_features.xdstp_based_config_singleton_subscriptions`
runtime guard which is currently disabled.
Testing: Added both unit-tests and integration tests
Docs Changes: N/A (until the entire xDS-TP based config sources support
is done).
Release Notes: N/A (until the entire xDS-TP based config sources support
is done).
Platform Specific Features: N/A

---------

Signed-off-by: Adi Suissa-Peleg <adip@google.com>