network: add additional methods on connection management interface (#40302)
## Description

This PR adds new virtual methods to the `Network::Connection` interface
to support socket ownership transfer which we need for the incoming
reverse connections feature. We are adding a new method `getSocket()`
that provides `const` access to connection socket reference.

---
**Commit Message:** network: add additional methods on connection
management interface
**Additional Description:** Adds new virtual methods to the
`Network::Connection` interface to support socket ownership transfer
which we need for the incoming reverse connections feature.
**Risk Level:** Low
**Testing:** Added Unit Tests
**Docs Changes:** N/A
**Release Notes:** N/A

---------

Signed-off-by: Rohit Agrawal <rohit.agrawal@databricks.com>