dns_resolver: add max_udp_channel_duration to allow periodic refresh of the UDP channel (#41085)
## Description

This PR added a new config called `max_udp_channel_duration` for the
c-ares DNS resolver which controls the lifetime of the UDP channel and
force re-initialize it after that time. We have seen quite a few cases
where kube-proxy updates leads to the existing UDP channel in a broken
state. Although the `max_udp_queries` are helpful, for regions were we
don't have enough traffic it still takes forever to recover.

This would help recovery by periodically refreshing the UDP channel in
cases mentioned above.

---

**Commit Message:** dns_resolver: add max_udp_channel_duration to allow
periodic refresh of the UDP channel
**Additional Description:** Added `max_udp_channel_duration`
configuration field to the c-ares DNS resolver which controls the
lifetime of the UDP channel and force re-initialize it after that time.
**Risk Level:** Low
**Testing:** Added Integration Tests
**Docs Changes:** Added
**Release Notes:** Added

Signed-off-by: Rohit Agrawal <rohit.agrawal@databricks.com>