async stream options: use shared pointer for pased retry policy (#41074)
This PR updated the StreamOptions. The shared pointer of RetryPolicy is
used rather than a reference to avoid potential lifetime problem when
the `send-and-forget` case is used for the async client.

This also fixed a potential bug at the Http::AsyncStreamImpl. At there,
if the parsing of proto retry policy is field, will result in core dump.

Risk Level: low.
Testing: n/a.
Docs Changes: n/a.
Release Notes: n/a.
Platform Specific Features: n/a.

---------

Signed-off-by: WangBaiping <wbphub@gmail.com>