ext_proc: add status_on_error option to control the status code returned (#41124)
## Description

This PR adds a new `status_on_error` in ExtProc similar to what we [have
for ExtAuthZ](https://arc.net/l/quote/zoibcqsy) which could be used to
control the status code returned back to the user in case of failures.
users can have some flow in which they expect a different status code to
be returned for failures instead of the hardcoded **500 Internal Server
Error**.

---

**Commit Message:** ext_proc: add status_on_error option to control the
status code returned
**Additional Description:** Added a new `status_on_error` in ExtProc to
control the HTTP status code returned back to the user in case of
failures.
**Risk Level:** Low
**Testing:** Added Integration Tests & Unit Tests
**Docs Changes:** Added
**Release Notes:** Added

---------

Signed-off-by: Rohit Agrawal <rohit.agrawal@databricks.com>