listener: Fix incorrect error status handling when failing to change netns (#40933)
When creating a listener socket in another network namespace, a nested
`absl::StatusOr<>` is returned. The outer status shows the result of the
attempt to switch network namespaces and the inner status is the result
of the creation of a listener socket.

This patch fixes a bug where we checked the inner status when we should
have been checking the outer status. This results in a segfault because
the outer status is dereferenced without first checking if it was OK.

We now check the correct results and additional comments have been added
to improve clarity of this code.

<!--
!!!ATTENTION!!!

If you are fixing *any* crash or *any* potential security issue, *do
not*
open a pull request in this repo. Please report the issue via emailing
envoy-security@googlegroups.com where the issue will be triaged
appropriately.
Thank you in advance for helping to keep Envoy secure.

!!!ATTENTION!!!

For an explanation of how to fill out the fields, please see the
relevant section
in
[PULL_REQUESTS.md](https://github.com/envoyproxy/envoy/blob/main/PULL_REQUESTS.md)
-->

Risk Level: Low
Testing: New unit test and fuzz test case
Docs Changes: n/a
Release Notes: Added
Platform Specific Features: Linux only

---------

Signed-off-by: Tony Allen <txallen@google.com>