Fix for broken FIPS builds (#41226)
[This PR](https://github.com/envoyproxy/envoy/pull/39728) moved the Go
dependency in FIPS builds to a Bazel `http_archive`. Once that was done,
Bazel's `filegroup` doesn't seem to be dealing so well with non-UTF-8
characters present in a file name in Go's test suite, causing the build
to fail.

```
#23 170.8 ERROR: /build/top/BUILD/envoy/build/bazel_root/base/external/boringssl_fips/BUILD.bazel:70:8: Executing genrule @@boringssl_fips//:build failed: error reading file '@@fips_go_linux_amd64//:test/fixedbugs/issue27836.dir/Þmain.go': /build/top/BUILD/envoy/build/bazel_root/base/external/fips_go_linux_amd64/test/fixedbugs/issue27836.dir/Þmain.go (No such file or directory)
#23 170.8 ERROR: /build/top/BUILD/envoy/build/bazel_root/base/external/boringssl_fips/BUILD.bazel:70:8: Executing genrule @@boringssl_fips//:build failed: 1 input file(s) are in error
#23 171.2 Target //distribution/binary:release failed to build
```

It's not fully understood what circumstances cause Bazel to behave like
that, but it seems reasonable to remove the `test` directory from the
exports. Once a bug is filed to the Bazel project I'll update the PR
with the issue number.

Signed-off-by: Gustavo <grnmeira@gmail.com>