ext_authz: check request header map count after applying mutations (#41122)
Commit Message: check request header map count after applying mutations
Additional Description:

This PR adds a new check to the ext_authz http filter. After request
header mutations are applied, this PR checks the number of headers in
the request header map and fails the downstream request if the total
number of headers is greater than the header limit for the request
header map.

There's more that can be done, but no need to do it in a single huge PR.
We can check the number of mutations before applying them. We can also
do the same things on the response path. This PR just adds the simplest
check for the request path.

Risk Level: low
Testing: unit / integration tested

---------

Signed-off-by: antoniovleonti <leonti@google.com>