router: Allow using another header for virtual host matching
This adds two things:

- A new RouteConfiguration option that specifies an alternative header
to be used instead of :authority/Host when selecting a virtual host
- A new early header mutation action that provides for regular
expression substituion into a different header.

Together, these allow for a "simplification" or "canonicalization"
step to be run on :authority, to reduce high cardinality portions of
the hostname to be replaced with a static string, without impacting
that :authority header that is sent upstream. This provides a way to
support multiple wildcards in a URL in a clean way.

Signed-off-by: Ryan Anderson <ryan@michonline.com>