[ext_proc] Introduce an extensible ProcessingRequestModifier interface (#41081)
Allow for custom extensions to change the processing request before it is sent on the wire.

Risk Level: Low
Testing: Added tests for both the no-op case, as well as a test
extension that exercises the extension and factory path.
Docs Changes:
Release Notes:
Platform Specific Features:

This is a re-work of https://github.com/envoyproxy/envoy/pull/40930.

---------

Signed-off-by: Misha Badov <mbadov@google.com>
Signed-off-by: mbadov <900737+mbadov@users.noreply.github.com>
Co-authored-by: phlax <phlax@users.noreply.github.com>