upstream: outlier detection for non 5xx codes (#39947)
Commit Message: outlier detection for non 5xx codes
Additional Description: 
Added to a cluster HTTP a specific option to define an http matcher.
When a response matches the defined matcher, the response is treated as
error, regardless of the status code. The error is reported to outlier
detection as 5xx code. If a response does not match the matcher it is
treated as success and forwarded to outlier detection as code 200.
Risk Level: Low (no change unless matcher is defined)
Testing: Added unit and integration tests
Docs Changes: yes
Release Notes: yes
Platform Specific Features: n/a
Fixes #18789

---------

Signed-off-by: Christoph Pakulski <paker8848@gmail.com>
Signed-off-by: Christoph Pakulski <christoph@tetrate.io>