grpc_transcoder additional test case (#41363)
Commit Message: grpc_transcoder additional test case
Additional Description: Since I introduced a bug in handling the
empty-message case for HttpBody streaming, and there was no test
coverage to prevent that bug, I thought it would be nice to have another
test case for what the empty-message case is supposed to do for another
type of request (I wasn't sure if this might be the cause of some other
flakiness we have - this test suggests it is not.)
Risk Level: None, test-only.
Testing: Yes it is.
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a

Signed-off-by: Raven Black <ravenblack@dropbox.com>