reverse_tunnels: add validation in the network filter (#41271)
## Description

This PR adds a `validation` to the Reverse Tunnel filter which could be
used to do validations on the incoming Node ID and Cluster ID values in
the reverse connection handshake. It's possible to use Filter State,
SNI, Certificate SAN, etc. to do these validations by configuring the
formatter. It's also possible to do validations on all or some of the
inputs.

---

**Commit Message:** reverse_tunnels: add validation in the network
filter
**Additional Description:** Adds `validation` to perform validations on
the incoming Node ID and Cluster ID from reverse connection handshake.
**Risk Level:** Low
**Testing:** Added Unit + Integration Tests
**Docs Changes:** N/A
**Release Notes:** N/A

---------

Signed-off-by: Rohit Agrawal <rohit.agrawal@databricks.com>