remove ravenblackx from deprecated cache ownership (#40985)
Commit Message: remove ravenblackx from deprecated cache ownership
Additional Description: With #41148 introducing a new cache filter
implementation that expects to be maintained, it no longer makes sense
to accept PRs touching the deprecated filter (other than to more firmly
deprecate it or move it to contrib where it doesn't need a maintainer).
Giving it the UNOWNED property is a gentle step towards making this
clearer, and signifying my intent to not review any PRs that aim to
"upkeep" the deprecated cache filter.

---------

Signed-off-by: Raven Black <ravenblack@dropbox.com>