quic: Only use the canonical suffix SRTT for TCP Delay duration
Currently, if the runtime guard `use_canonical_suffix_for_srtt` is true,
getSrtt() always checks if there is a SRTT estimate for any origin with
the same canonical suffix. However, getSrtt() is used for setting the
initial QuicConfig RTT estimate when creating a QUIC connection. In that
case, we shouldn't use the canonical suffix match for fetching the SRTT
value. The canonical suffix should only be used when calling getSrtt()
to determine the delay before attempting a TCP connection after
initiating a QUIC connection.

Signed-off-by: Ali Beyad <abeyad@google.com>