http: add stats for stateful session filter (#41236)
## Description

Today when the stateful session filter is running in the `strict=false`
mode it's hard to understand the cases where it falls back on the
default load balancing after not getting a routable destination.

In this PR we are trying to add stats on how many requests were failed
open and failed closed to improve the observability.

---

**Commit Message:** http: add stats for stateful session filter
**Additional Description:** Added stats in stateful session filter on
how many requests were failed open and failed closed to improve the
observability.
**Risk Level:** Low
**Testing:** Added Unit + Integration Tests
**Docs Changes:** Added
**Release Notes:** Added

---------

Signed-off-by: Rohit Agrawal <rohit.agrawal@databricks.com>