mobile: Re-enable xDS in Envoy Mobile (#41464)
This change re-enables xDS functionality in Envoy Mobile, after it was
removed in https://github.com/envoyproxy/envoy/pull/35017 and
https://github.com/envoyproxy/envoy/pull/35031.

xDS will now be built into Envoy Mobile, unless being built for the
Android or iOS platforms, in which case xDS will be disabled (for binary
size considerations on mobile devices).

This change also updates several places in the xDS implementation that
use `throw` and `catch`, since Envoy Mobile is built without exceptions
and thus, fails to compile with the current `throw` and `catch` blocks.
Instead, we use throwEnvoyExceptionOrPanic() and the CATCH() macros.

---------

Signed-off-by: Ali Beyad <abeyad@google.com>