mobile: re-enable xDS in Envoy Mobile (#41518)
https://github.com/envoyproxy/envoy/pull/41464 was reverted due to
failing tests on CI. The issue was that a new mobile workflow was added
on CI using custom Bazel flags to exercise the xDS build
(`--define=envoy_mobile_xds=enabled`) but when I ran the test locally, I
forgot to add those flags so they never actually ran.

I was able to reproduce the CI failure, fixed the tests, and now the
tests pass.

This change restores the original PR and adds fixes for the failing test
code.

---------

Signed-off-by: Ali Beyad <abeyad@google.com>