Add NE comparator for accesslog ComparisonFilter (#41572)
Commit Message: Add NE comparator for accesslog ComparisonFilter
Additional Description: The absence of this comparator has been a
recurring annoyance for us, with variations of `Or{{LE: val-1}, {GE,
val+1}}` scattered in our config. It finally reached the point where we
want to configure an exclusion of *two* status codes, which is
impossible with these comparators and no "not" compound matcher. Turns
out this addition is so trivial that I should have done it years ago and
saved all that pain.
Risk Level: Very low, only a new path that will only be taken with a
corresponding new config.
Testing: Yes.
Docs Changes: In the proto docgen.
Release Notes: Added support for not-equal operator for access log
filter rules, in `ComparisonFilter`
Platform Specific Features: n/a

---------

Signed-off-by: Raven Black <ravenblack@dropbox.com>