router: Allow using another header for virtual host matching (#41280)
This adds a new RouteConfiguration option that specifies an alternative
header to be used instead of :authority/Host when selecting a virtual
host

This allows for a "simplification" or "canonicalization" step to be run
on :authority, to reduce high cardinality portions of the hostname to be
replaced with a static string, without impacting that :authority header
that is sent upstream. This provides a way to support multiple wildcards
in a URL in a clean way.

Commit Message: router: Allow using another header for virtual host
matching
Additional Description:
This adds a new RouteConfiguration option that specifies an alternative
header to be used instead of :authority/Host when selecting a virtual
host

This allows for a "simplification" or "canonicalization" step to be run
on :authority, to reduce high cardinality portions of the hostname to be
replaced with a static string, without impacting that :authority header
that is sent upstream. This provides a way to support multiple wildcards
in a URL in a clean way.

Risk Level:
Testing: added new tests
Docs Changes: Updated api definition
Release Notes: Updated changelog
Platform Specific Features:
[Optional Runtime guard:]
[Optional Fixes #Issue]
[Optional Fixes commit #PR or SHA]
[Optional Deprecated:]
[Optional [API
Considerations](https://github.com/envoyproxy/envoy/blob/main/api/review_checklist.md):]

---------

Signed-off-by: Ryan Anderson <ryan@michonline.com>