quic: fix proof verifier factory registration by using ProofSourceConfig
The EnvoyQuicServerProofVerifierFactory was creating a TypedExtensionConfig
with the wrong proto message type. The factory name was correctly set to
'envoy.quic.server.proof_verifier.filter_chain', but the typed_config was
being packed with ProofSourceConfig proto with the wrong type URL.

This caused a factory lookup failure in Envoy Mobile's client_integration_test
with the error: 'Didn't find a registered implementation for
'envoy.quic.server.proof_verifier.filter_chain' with type URL:
'envoy.extensions.quic.proof_source.v3.ProofSourceConfig''.

The fix is to consistently use ProofSourceConfig (which is an empty proto)
for both the factory's createEmptyConfigProto() method and when creating
the default TypedExtensionConfig in active_quic_listener.cc. Since
ProofSourceConfig is empty and serves a similar purpose (QUIC server-side
configuration), this avoids the need to create and track a new proto file.

Signed-off-by: Rohit Agrawal <rohit.agrawal@databricks.com>