formatter: pure renaming of the formatter method and the context (#41698)
Commit Message: formatter: pure renaming of the formatter method and the
context
Additional Description:

1. Updated the Formatter::HttpFormatterContext to Formatter::Context
because it also be used for non HTTP protocol now.
2. Update the `formatWithContext/formatValueWithContext` to
`format/formatValue`. The latter is its original name. We use the former
was used for a while when we trying to support multiple protocols in the
substitution formatter.

Risk Level: low.
Testing: n/a.
Docs Changes: n/a.
Release Notes: n/a.
Platform Specific Features: n/a.

---------

Signed-off-by: wbpcode <wbphub@gmail.com>