socket: include network namespace in proto translation
Summary:
Admin `/listeners` endpoint does not return the network namespace path of
the listener sockets. This pull request updates the utilities used by
listener handler to include network namespace path in the proto address
returned in the response. This is used in production used for setting up
sockets per network namespace and monitoring.

Signed-off-by: Ronak Jain <ronakjainc@gmail.com>