http: Support for Proxy Protocol based on request downstream connection information.
Adds a new option `add_proxy_protocol_request_state` which if set to True will propagate the downstream connection information to the proxy protocol filter state. This enables propagating the client IP based on x-forwarded-for headers. This is useful in environments that encapsulate TCP connections over an HTTP/2 CONNECT stream.

Signed-off-by: Braden Bassingthwaite <bbassingthwaite@digitalocean.com>