ext_proc: closing the gRPC stream ASAP once no more external processing needed (#41425)
This is to address the issue:
https://github.com/envoyproxy/envoy/issues/41764, i.e, close the gRPC
stream once no further external processing needed. It is also partially
discussed in https://github.com/envoyproxy/envoy/issues/37088.

Currently the ext_proc gRPC stream is opened when the 1st
ProcessingRequest is sent to the ext_proc server. And it is closed
during ext_proc filter destruction. This is wasting resource on both
Envoy and ext_proc server side. For example, if envoy is configured to
only send request headers, the gRPC stream is left open until all the
way to the response is processed.

This PR is trying to close the ext_proc gRPC stream once Envoy detects
no more external processing needed.

---------

Signed-off-by: Yanjun Xiang <yanjunxiang@google.com>