rate_limit_quota filter: Enable preview-processing of matched buckets set to `keep_matching` (#41849)
Add support for the `keep_matching` field in MatchTrees.
This setting will mark a bucket as having been matched, but not chosen
as the final allow/deny deciding bucket.
A field that is skipped via `keep_matching` is treated as if it was in a
preview/darklaunch mode. This means that the first skipped bucket has
its usage recorded & allow/deny decision reached, as if it were matched
as the decision-making bucket, but the final result is ignored instead
of enforced. This allows for changes to the matchers to be tested
without affecting live traffic.

Additional Description: This also includes a refactor of the integration
tests to cleanup a lot of code repetition while I was adding options for
preview-mode testing.

Risk Level:
Testing: unit & manual
Docs Changes:
Release Notes:
Platform Specific Features:

Signed-off-by: Brian Surber <bsurber@google.com>