Make overloaded-virtual an error rather than a warning (#41898)
We keep re-introducing this warning into envoy
which sometimes provokes huge extremely noisy build logs (in one case it
was enough to overwhelm the CI interface to the point that you couldn't
even view the logs in the web UI). Even though it is the case that gcc's
warning threshold for this issue is a bit oversensitive, it's not that
hard to *not* provoke the warning/error, so it seems quite reasonable to
treat it in the default way, as an error, and thereby prevent us from
accidentally reintroducing spammy warnings.

A reasonable alternative would be to tag gcc builds
`-Wno-overloaded-virtual` rather than `-Wno-error=overloaded-virtual`
which would suppress the warnings as well, but the warnings aren't
*wrong*, in that you can cause yourself grief with this hidden overload
issue, even if, like the old comment said, it can't cause misbehavior.
Risk Level: No behavior change.
Testing: No behavior change.
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a

---------

Signed-off-by: Raven Black <ravenblack@dropbox.com>