router: added request_mirror_policies to cluster (#41834)
## Description

This PR adds `request_mirror_policies` to the cluster for enabling
request mirroring to be configured at the cluster level. When
configured, all requests routed to the cluster will be mirrored
according to the specified policies.

It's very helpful to have this for ExtProc, ExtAuthZ, RateLimit, etc.

Fix https://github.com/envoyproxy/envoy/issues/21131

---

**Commit Message:** router: added request_mirror_policies to cluster
**Additional Description:** Adds `request_mirror_policies` to the
cluster for enabling request mirroring to be configured at the cluster
level.
**Risk Level:** Low
**Testing:** Added Unit + Integration Tests
**Docs Changes:** Added
**Release Notes:** Added

---------

Signed-off-by: Rohit Agrawal <rohit.agrawal@databricks.com>
Signed-off-by: yanavlasov <yavlasov@google.com>
Co-authored-by: yanavlasov <yavlasov@google.com>