ext_authz: add response fuzzing to ext_authz grpc fuzz test (#41895)
Commit Message: add response fuzzing to ext_authz grpc fuzz test
Additional Description:

Add response fuzzing in the simplest way possible to the ext authz grpc
fuzzer.

This fuzzer now, regardless of what happened on the request path, will
also invoke the response path. This is not the "highest fidelity" way to
do it but it is the simplest and I ran the test overnight with zero
crashes, so I think it's fine for improving fuzz coverage for the ext
authz filter.

Risk Level: none
Testing: fuzz improvement

Signed-off-by: antoniovleonti <leonti@google.com>