fix mcp filter fuzz case (#42067)
Commit Message: fix mcp filter fuzz case
Additional Description:

This PR fixes a fuzz case that causes the mcp filter to call
addDecodedData followed by decodingBuffer. In the real decoder callbacks
a new buffer is created even if the data is empty. This is not simulated
in the expectations. This pr fixes the expectations to be more faithful
to the actual implementation.

Risk Level: none
Testing: fuzz fix

Signed-off-by: antoniovleonti <leonti@google.com>